Added lots of new stuff: See additional comments
Viewing of books now works in real time.
Added code to view games, employees, and customers in real time.
Code to view movies and music still needs to be added.

Added forms to add employees, customers, music, games, and movies.
None of the SQL code to add these to the database are done but the GUIs
for the most part should be.

The database still needs some work to reflect how we are storing data,
which has changed significantly since the database was made.